move class to proper package
