+++ app/src/main/java/com/example/mymusicapp/service/MusicService.java
@@ -1,4 +1,4 @@
-package com.example.mymusicapp;
+package com.example.mymusicapp.service;
 
 import android.app.Service;
 import android.content.ContentUris;
@@ -7,13 +7,11 @@ import android.media.AudioManager;
 import android.media.MediaPlayer;
 import android.net.Uri;
 import android.os.Binder;
-import android.os.Build;
 import android.os.IBinder;
 import android.os.PowerManager;
 import android.util.Log;
 
 import androidx.annotation.Nullable;
-import androidx.annotation.RequiresApi;
 
 import com.example.mymusicapp.entity.Song;
